extract function as a package
